requirements.txt
pandas
numpy
Pillow
scikit-learn
scikit-image
scipy
tqdm
pymupdf
aiofiles
